Fixed tabs-mpv not killing mpv after exiting
I'm pretty sure this started happening after I changed my /bin/sh
symlink to Dash. Bash seems to run the EXIT trap when the proccess is
killed using kill but Dash does not. Changing the trap from EXIT to TERM
(which is what kill is actually sending to the script) seems to fix the
issue.